MoreUtils.xs: Fix RT#80385 - part() output references input in XS, not in PP
Fix RT#80385 after approving and applying provided fixes.

The XS version of part() appears to create scalars in its output by making
aliases to values in its input stream. Subsequent changes to either the
input or the output values, end up affecting the other. This behavior is not
present in the PP implementation.

Signed-off-by: Jens Rehsack <sno@netbsd.org>