Changes: prepare for release
Note changes made and bump version for a recommend -> prereq_pm change
requested quickly by Johan Lindstrom and Andreas Koenig in RT#122875.
Both strongly voted for a revert to 0.429 which would hurt all users
of new functions, so the only reasonable offer is dealing better with
some tool issues.

It was an error to split PP and XS to get rid of Config::AutoConf in
base distribution - but the error has been made and is in the world,
so a rollback is not possible. Discussing this is insane - looking
forward is the only option.

Signed-off-by: Jens Rehsack <sno@netbsd.org>